update: fullfilled 3 user stories
Requirements

user story 1:
As a user, I want to see a table with two elements (three-way segmented control, color wheel) when I open the app, so that I can choose a color.
Acceptance Criteria:
1. App opens and runs stable.
2. No memory leaks.
3. Scrollable content.

user story 2:
As a user, I want to be able to use a color wheel to select a color and see the selected color in a tinted circle icon, so that I can easily and quickly choose my desired color visually.
Acceptance Criteria:
1. Color wheel properties
a. The color wheel’s hue value is distributed radially over 360 degrees.
b. The color wheel loses saturation towards the middle, where the saturation is 0%.
c. The color wheel does not have the ability to change the brightness of the color, the value is 100%.
2. Layout
a. The color wheel should be resizable with screen width, with fixed aspect ratio and a constant margin of 24 towards the leading and trailing edge of the screen.
3. Programmatically create the visual using a radial gradient, or, if too hard, use an image.
4. Moving the color wheel handle updates the tint color of the selected circle icon in the three- way segmented control.

user story 3:
As a user, I want to select one of three tinted circle icons and see the color in the right position on a color wheel, so that I can modify the color easily by moving to neighboring colors on the wheel.
Acceptance Criteria:
1. Tap on a control.
a. highlights the control area background.
b. Moves the color wheel handle to the correct position on the color wheel.